# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Web Dev Blog Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/slug/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      summary: Get a category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid credentials
...
  /auth/change-password:
    post:
      operationId: changeAdminPassword
      tags: [auth]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
  /auth/change-username:
    post:
      operationId: changeAdminUsername
      tags: [auth]
      summary: Change admin username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUsernameInput"
      responses:
        "200":
          description: Username changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeUsernameResponse"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
  /auth/forgot-password:
    post:
      operationId: forgotAdminPassword
      tags: [auth]
      summary: Reset admin password using recovery token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, postCount, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        postCount: { type: integer }
        createdAt: { type: string }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string }
        description: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string }
        description: { type: string }
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        username: { type: ["string", "null"] }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 6 }
    ChangeUsernameInput:
      type: object
      required: [currentPassword, newUsername]
      properties:
        currentPassword: { type: string }
        newUsername: { type: string, minLength: 1 }
    ChangeUsernameResponse:
      type: object
      required: [success, username]
      properties:
        success: { type: boolean }
        username: { type: string }
    ForgotPasswordInput:
      type: object
      required: [username, recoveryToken, newPassword]
      properties:
        username: { type: string }
        recoveryToken: { type: string }
        newPassword: { type: string, minLength: 6 }
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...